backend/pyproject.toml
[project]
name = "vectorspace-backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "chromadb>=1.0.0",
    "email-validator>=2.0.0",
    "fastapi>=0.115.9",
    "greenlet>=3.2.3",
    "openai>=1.88.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sentence-transformers>=4.1.0",
    "sqlalchemy>=2.0.41",
    "uvicorn[standard]>=0.34.3",
]

[dependency-groups]
dev = [
    "faker>=37.4.0",
    "httpx>=0.28.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-bdd>=8.1.0",
    "pytest-cov>=6.2.1",
    "pytest-xdist>=3.7.0",
]
